+++ src/lib.rs
@@ -12,14 +12,14 @@
 //! ## Associative binary operation
 //!
 //! ```
-//! use diceprop::{infix_fun_2, props, FateVarExt};
+//! use diceprop::{infix_fun_2, props, FateVarExt, Set};
 //! use dicetest::prelude::*;
 //!
 //! #[test]
 //! fn add_is_associative_for_small_f32() {
 //!     Dicetest::repeatedly().run(|mut fate| {
-//!         let small_f32_die = dice::f32(-100.0..=100.0);
-//!         let var = fate.roll_var("f32 ∩ [-100,100]", ["x", "y", "z"], small_f32_die);
+//!         let set = Set::new("f32 ∩ [-100,100]", dice::f32(-100.0..=100.0));
+//!         let var = fate.roll_var(["x", "y", "z"], set);
 //!         let add = infix_fun_2("+", |x, y| x + y);
 //!         props::binop::associative(var, add);
 //!     })
@@ -57,14 +57,14 @@
 //! ## Left inverse function
 //!
 //! ```
-//! use diceprop::{fun_1, postfix_fun_1, props, FateVarExt};
+//! use diceprop::{fun_1, postfix_fun_1, props, FateVarExt, Set};
 //! use dicetest::prelude::*;
 //!
 //! #[test]
 //! fn sqrt_is_left_inverse_of_sq_for_non_negative_f32() {
 //!     Dicetest::repeatedly().run(|mut fate| {
-//!         let non_negative_f32_die = dice::f32(0.0..);
-//!         let var = fate.roll_var("f32 ∩ [0,+∞]", ["x"], non_negative_f32_die);
+//!         let set = Set::new("f32 ∩ [0,+∞]", dice::f32(0.0..));
+//!         let var = fate.roll_var(["x"], set);
 //!         let sq = postfix_fun_1("²", |x| x * x);
 //!         let sqrt = fun_1("√", |x: f32| x.sqrt());
 //!         props::fun::left_inverse(var, sq, sqrt);
@@ -99,14 +99,14 @@
 //! ## Partial order
 //!
 //! ```
-//! use diceprop::{infix_fun_2, props, FateVarExt};
+//! use diceprop::{infix_fun_2, props, FateVarExt, Set};
 //! use dicetest::prelude::*;
 //!
 //! #[test]
 //! fn gt_is_partial_order_for_any_f32() {
 //!     Dicetest::repeatedly().run(|mut fate| {
-//!         let any_f32_die = dice::any_f32();
-//!         let var = fate.roll_var("f32", ["x", "y", "z"], any_f32_die);
+//!         let set = Set::new("f32", dice::any_f32());
+//!         let var = fate.roll_var(["x", "y", "z"], set);
 //!         let gt = infix_fun_2("≤", |x, y| x <= y);
 //!         props::binrel::partial_order(var, gt);
 //!     })
@@ -142,6 +142,9 @@ pub use eval::Eval;
 
 mod elem;
 pub use elem::{elem, Elem};
+
+mod set;
+pub use set::Set;
 
 mod var;
 pub use var::{FateVarExt, Var};